+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 * Fixed cursor loading on Clickhouse
+* Improved batch block flush logic to flush after a certain number of blocks instead of taking by modulo
 
 ## v4.5.0